+++ src/app/cart/item/item.component.ts
@@ -4,7 +4,7 @@ import {CartItem} from '../item.model';
 @Component({
     selector: 'app-cart-item',
     templateUrl: './item.component.html',
-    styleUrls: ['./item.component.css']
+    styleUrls: ['./item.component.scss']
 })
 
 export class ItemComponent {