+++ src/app/cart/list/cart.component.ts
@@ -7,7 +7,7 @@ import {CommunicatorService} from '../../communicator.service';
 @Component({
     selector: 'app-cart',
     templateUrl: './cart.component.html',
-    styleUrls: ['./cart.component.css']
+    styleUrls: ['./cart.component.scss']
 })
 export class CartComponent implements OnInit, OnDestroy {
     private sub: Subscription;